Test conversions on image test file
