+++ podoc/ast/test_files/image.json
@@ -60,11 +60,24 @@
                         ],
                         [
                             {
-                                "c": [
-                                    "html",
-                                    "<matplotlib.figure.Figure at 0x7fe727fed7b8>"
-                                ],
-                                "t": "RawInline"
+                                "c": "<matplotlib.figure.Figure",
+                                "t": "Str"
+                            },
+                            {
+                                "c": [],
+                                "t": "Space"
+                            },
+                            {
+                                "c": "at",
+                                "t": "Str"
+                            },
+                            {
+                                "c": [],
+                                "t": "Space"
+                            },
+                            {
+                                "c": "0x7fe727fed7b8>",
+                                "t": "Str"
                             }
                         ],
                         [
